+++ crates/tools/lib/Cargo.toml
@@ -1,10 +1,8 @@
 [package]
-name = "tool_bindings"
+name = "tool_lib"
 version = "0.0.0"
 edition = "2018"
 publish = false
 
 [dependencies]
 metadata = { package = "windows-metadata", path = "../../libs/metadata" }
-bindgen = { package = "windows-bindgen", path = "../../libs/bindgen" }
-lib = { package = "tool_lib", path = "../lib" }
