+++ crates/tools/sys/src/main.rs
@@ -94,10 +94,10 @@ deprecated = []
 }
 
 fn gen_tree(reader: &metadata::reader::Reader, output: &std::path::Path, tree: &metadata::reader::Tree) {
+    println!("{}", tree.namespace);
     let mut path = std::path::PathBuf::from(output);
     path.push(tree.namespace.replace('.', "/"));
     std::fs::create_dir_all(&path).unwrap();
-    path.push("mod.rs");
 
     let mut gen = bindgen::Gen::new(reader);
     gen.namespace = tree.namespace;
@@ -105,19 +105,6 @@ fn gen_tree(reader: &metadata::reader::Reader, output: &std::path::Path, tree: &
     gen.cfg = true;
     gen.doc = true;
     let mut tokens = bindgen::namespace(&gen, tree);
-
-    let mut child = std::process::Command::new("rustfmt").stdin(std::process::Stdio::piped()).stdout(std::process::Stdio::piped()).stderr(std::process::Stdio::null()).spawn().expect("Failed to spawn `rustfmt`");
-    let mut stdin = child.stdin.take().expect("Failed to open stdin");
-    stdin.write_all(tokens.as_bytes()).unwrap();
-    drop(stdin);
-    let output = child.wait_with_output().unwrap();
-
-    if output.status.success() {
-        println!("{}", tree.namespace);
-        tokens = String::from_utf8(output.stdout).expect("Failed to parse UTF-8");
-    } else {
-        println!("** {} - rustfmt failed", tree.namespace);
-    }
-
-    std::fs::write(&path, tokens).unwrap();
+    lib::rustfmt(tree.namespace, &mut tokens);
+    std::fs::write(path.join("mod.rs"), tokens).unwrap();
 }
