+++ crates/tools/windows/src/main.rs
@@ -113,24 +113,10 @@ fn gen_tree(reader: &metadata::reader::Reader, output: &std::path::Path, tree: &
     gen.min_xaml = true;
     let mut tokens = bindgen::namespace(&gen, tree);
     tokens.push_str(r#"#[cfg(feature = "implement")] ::core::include!("impl.rs");"#);
-    fmt_tokens(tree.namespace, &mut tokens);
+    lib::rustfmt(tree.namespace, &mut tokens);
     std::fs::write(path.join("mod.rs"), tokens).unwrap();
 
     let mut tokens = bindgen::namespace_impl(&gen, tree);
-    fmt_tokens(tree.namespace, &mut tokens);
+    lib::rustfmt(tree.namespace, &mut tokens);
     std::fs::write(path.join("impl.rs"), tokens).unwrap();
-}
-
-fn fmt_tokens(namespace: &str, tokens: &mut String) {
-    let mut child = std::process::Command::new("rustfmt").stdin(std::process::Stdio::piped()).stdout(std::process::Stdio::piped()).stderr(std::process::Stdio::null()).spawn().expect("Failed to spawn `rustfmt`");
-    let mut stdin = child.stdin.take().expect("Failed to open stdin");
-    stdin.write_all(tokens.as_bytes()).unwrap();
-    drop(stdin);
-    let output = child.wait_with_output().unwrap();
-
-    if output.status.success() {
-        *tokens = String::from_utf8(output.stdout).expect("Failed to parse UTF-8");
-    } else {
-        println!("** {} - rustfmt failed", namespace);
-    }
 }
